requirements.txt
annotated-types==0.7.0
colorama==0.4.6
mss==10.0.0
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
pytesseract==0.3.13
typing_extensions==4.12.2
keyboard
bettercam
fuzzywuzzy
python-Levenshtein
